requirements.txt
vanna[sqlite,openai]>=2.0.0
fastapi
uvicorn[standard]
plotly
pandas
python-dotenv
openai
groq
pydantic
